fix(core): 🐛 fix Windows cross-compilation type errors (#219)

+++ src-tauri/src/core/gateway_supervisor.rs
@@ -292,7 +292,7 @@ fn is_process_alive(pid: u32) -> bool {
         let mut exit_code: u32 = 0;
         let result = GetExitCodeProcess(handle, &mut exit_code);
         CloseHandle(handle);
-        result != 0 && exit_code == STILL_ACTIVE
+        result != 0 && exit_code == STILL_ACTIVE as u32
     }
 }
